06/16

+++ SpringBoard/src/main/java/com/ict/persistence/PageMaker.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class PageMaker {
 
-	private int totalBoard;
+	private Long totalBoard;
 	private int startPage;
 	private int endPage;
 	private boolean prev;
@@ -32,7 +32,7 @@ public class PageMaker {
 		this.next = this.endPage * cri.getNumber() >= this.totalBoard ? false : true;
 	}
 	
-	public void setTotalBoard(int totalBoard) {
+	public void setTotalBoard(Long totalBoard) {
 		this.totalBoard = totalBoard;
 		calcData();
 	}
